Complete refactor
