+++ youtui/src/app/ui.rs
@@ -9,7 +9,9 @@ use super::structures::*;
 use super::view::Scrollable;
 use super::AppCallback;
 use crate::async_rodio_sink::{SeekDirection, VolumeUpdate};
-use crate::config::{AppAction, Config, KeyEnum, KeyEnumKey};
+use crate::config::keybinds::{KeyEnum, KeyEnumKey};
+use crate::config::Config;
+use action::AppAction;
 use async_callback_manager::{AsyncTask, Constraint};
 use crossterm::event::{Event, KeyEvent};
 use ratatui::widgets::TableState;
